blockchain not in working state. Rounds are cancelled and the program is very slow
In main, the node tries to get the data from the writer api but otherwise runs with its local files. In protoengine, writers check for updates to config, connect to everyone, and add missing blocks before starting up. Every 20 rounds the nodes update and join again to the writer set. Before every round starts, the nodes check for changes to the active nodes list.