nodes are able to keep going with block length of > 0, when view-change is not required, and nodes have the same round.
Yet to add verified announce.
Fails on zero blocks appended.
Nodes revert the latest block on cancellation.
The fault is designed for everyone knowing that a node turned off.
tested with three nodes, one of which is down.
Faulty node is unable to catch-up in current state or join again after going down.